Okay, Creator/Non-Creator mode should Do The Right Thing now.
Sort of tightened that up and made things more explicit too; I think
BLCMM had sort of been relying on some CWD fiddling for some of that,
in the past.  Things which open files now should be mostly doing so
via explicit paths.

Also merged BLCMMUtilities into Utilities because that was too
confusing.