More font-size fixups: OE Object Browser instructions resize fix
It turns out that with *only* leaf attributes visible in our trees,
changing the font size will *not* adjust the spacing between each line,
even though it *does* work when there's actual trees to display.  Super
weird.  The easy fix here is to just show the root folder (and give it
the label "Instructions" so it doesn't look weird).  Good enough!

This continues the work on #28, though more work is needed.